Reverting some chages to keep records
This commit reverts a number of changes to the shell component to the state it was in a number of commits before. A number of changes from this template have already been integrated in the main upstream branch via separate commits / pull requests. It is not clear whether the remaining changes (primarily navbar menu buttons) will be implemented in the upstream code base. Note, another not propagated change in this template is the reset of the min-width attribute on all navbar menu elements. This approach to remove the extra white spaces on the right of the dropdown submenus prooved to be buggy, particularly when both icons and text labels are displayed on menu elements. It appears to work ok when only one of the two is used on the whole submenu.